+++ src/main/kotlin/org/wfanet/measurement/reporting/deploy/postgres/PostgresMeasurementsService.kt
@@ -22,7 +22,6 @@ import org.wfanet.measurement.internal.reporting.Measurement
 import org.wfanet.measurement.internal.reporting.MeasurementsGrpcKt.MeasurementsCoroutineImplBase
 import org.wfanet.measurement.internal.reporting.SetMeasurementFailureRequest
 import org.wfanet.measurement.internal.reporting.SetMeasurementResultRequest
-import org.wfanet.measurement.reporting.deploy.postgres.PostgresSerializable.withSerializableErrorRetries
 import org.wfanet.measurement.reporting.deploy.postgres.readers.MeasurementReader
 import org.wfanet.measurement.reporting.deploy.postgres.writers.CreateMeasurement
 import org.wfanet.measurement.reporting.deploy.postgres.writers.SetMeasurementFailure
@@ -44,7 +43,7 @@ class PostgresMeasurementsService(
 
   override suspend fun getMeasurement(request: GetMeasurementRequest): Measurement {
     return try {
-      withSerializableErrorRetries {
+      SerializableErrors.retrying {
         MeasurementReader()
           .readMeasurementByReferenceIds(
             client.singleUse(),
