+++ src/main/kotlin/org/wfanet/measurement/reporting/deploy/postgres/PostgresReportingSetsService.kt
@@ -24,7 +24,7 @@ import org.wfanet.measurement.internal.reporting.GetReportingSetRequest
 import org.wfanet.measurement.internal.reporting.ReportingSet
 import org.wfanet.measurement.internal.reporting.ReportingSetsGrpcKt.ReportingSetsCoroutineImplBase
 import org.wfanet.measurement.internal.reporting.StreamReportingSetsRequest
-import org.wfanet.measurement.reporting.deploy.postgres.PostgresSerializable.withSerializableErrorRetries
+import org.wfanet.measurement.reporting.deploy.postgres.SerializableErrors.withSerializableErrorRetries
 import org.wfanet.measurement.reporting.deploy.postgres.readers.ReportingSetReader
 import org.wfanet.measurement.reporting.deploy.postgres.writers.CreateReportingSet
 import org.wfanet.measurement.reporting.service.internal.ReportingSetAlreadyExistsException
@@ -46,7 +46,7 @@ class PostgresReportingSetsService(
 
   override suspend fun getReportingSet(request: GetReportingSetRequest): ReportingSet {
     return try {
-      withSerializableErrorRetries {
+      SerializableErrors.retrying {
         ReportingSetReader()
           .readReportingSetByExternalId(
             client.singleUse(),
